messenger/requirements.txt
Flask
Flask-SocketIO
Flask-Login
python-dotenv
google-cloud-language
requests
gitpython